add save and acquisition function to trigger call back function
added functionallity to the trigger callback function. This is where the
session data is processed now. additionally i debugged some of the other
functions and the main script runs now either with a trigger and with an
immediate call. Also I added 2 new functions, for simulation mode and to
run when triggered. The info etc in headers need to be changed in next
upload!